+++ u8pool/tests/iterator_tests.rs
@@ -35,19 +35,21 @@ fn test_iterator_populated_vector() {
 }
 
 #[test]
-fn test_iterator_consumed_completely() {
+fn test_iterator_single_item() {
     let mut buffer = [0u8; 600];
     let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
 
-    u8pool.push(b"a").unwrap();
-    u8pool.push(b"b").unwrap();
+    u8pool.push(b"only").unwrap();
 
-    let collected: Vec<_> = u8pool.into_iter().collect();
-    assert_eq!(collected, vec![&b"a"[..], &b"b"[..]]);
+    let mut iter = u8pool.into_iter();
+    assert_eq!(iter.size_hint(), (1, Some(1)));
+    assert_eq!(iter.next(), Some(&b"only"[..]));
+    assert_eq!(iter.size_hint(), (0, Some(0)));
+    assert_eq!(iter.next(), None);
 }
 
 #[test]
-fn test_iterator_partial_iteration() {
+fn test_iterator_size_hint() {
     let mut buffer = [0u8; 600];
     let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
 
@@ -56,57 +58,90 @@ fn test_iterator_partial_iteration() {
     u8pool.push(b"third").unwrap();
 
     let mut iter = u8pool.into_iter();
-    assert_eq!(iter.next(), Some(&b"first"[..]));
-    assert_eq!(iter.next(), Some(&b"second"[..]));
-    // Iterator should still work after partial consumption
+    assert_eq!(iter.size_hint(), (3, Some(3)));
+
+    iter.next();
+    assert_eq!(iter.size_hint(), (2, Some(2)));
+
+    iter.next();
     assert_eq!(iter.size_hint(), (1, Some(1)));
-    assert_eq!(iter.next(), Some(&b"third"[..]));
-    assert_eq!(iter.next(), None);
+
+    iter.next();
+    assert_eq!(iter.size_hint(), (0, Some(0)));
 }
 
 #[test]
-fn test_iterator_lifetime_correctness() {
+fn test_reverse_iterator_populated_vector() {
     let mut buffer = [0u8; 600];
     let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
 
-    u8pool.push(b"data").unwrap();
+    u8pool.push(b"first").unwrap();
+    u8pool.push(b"second").unwrap();
+    u8pool.push(b"third").unwrap();
 
-    // Test that iterator can be created and used
-    {
-        let iter = u8pool.into_iter();
-        let first = iter.take(1).next().unwrap();
-        assert_eq!(first, b"data");
+    let items: Vec<_> = u8pool.iter_rev().collect();
+    assert_eq!(items.len(), 3);
+    assert_eq!(items[0], b"third");
+    assert_eq!(items[1], b"second");
+    assert_eq!(items[2], b"first");
+}
+
+#[test]
+fn test_reverse_iterator_empty_vector() {
+    let mut buffer = [0u8; 600];
+    let u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    let items: Vec<_> = u8pool.iter_rev().collect();
+    assert_eq!(items.len(), 0);
+}
+
+#[test]
+fn test_reverse_iterator_single_item() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    u8pool.push(b"only").unwrap();
+
+    let items: Vec<_> = u8pool.iter_rev().collect();
+    assert_eq!(items.len(), 1);
+    assert_eq!(items[0], b"only");
+}
+
+#[test]
+fn test_reverse_iterator_size_hint() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    u8pool.push(b"first").unwrap();
+    u8pool.push(b"second").unwrap();
+    u8pool.push(b"third").unwrap();
+
+    let mut iter = u8pool.iter_rev();
+    assert_eq!(iter.size_hint(), (3, Some(3)));
+
+    iter.next();
+    assert_eq!(iter.size_hint(), (2, Some(2)));
+
+    iter.next();
+    assert_eq!(iter.size_hint(), (1, Some(1)));
+
+    iter.next();
+    assert_eq!(iter.size_hint(), (0, Some(0)));
+}
+
+#[test]
+fn test_reverse_iterator_compare_with_forward() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    let items: &[&[u8]] = &[b"alpha", b"beta", b"gamma", b"delta"];
+    for item in items {
+        u8pool.push(item).unwrap();
     }
 
-    // U8Pool should still be usable after iterator is dropped
-    assert_eq!(u8pool.len(), 1);
-    assert_eq!(u8pool.get(0).unwrap(), b"data");
-}
+    let forward: Vec<_> = u8pool.iter().collect();
+    let mut reverse: Vec<_> = u8pool.iter_rev().collect();
+    reverse.reverse(); // Reverse it back to compare
 
-#[test]
-fn test_for_loop_syntax() {
-    let mut buffer = [0u8; 600];
-    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
-
-    u8pool.push(b"hello").unwrap();
-    u8pool.push(b"world").unwrap();
-
-    let mut results = Vec::new();
-    for slice in &u8pool {
-        results.push(slice);
-    }
-
-    assert_eq!(results, vec![&b"hello"[..], &b"world"[..]]);
-}
-
-#[test]
-fn test_iter_method() {
-    let mut buffer = [0u8; 600];
-    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
-
-    u8pool.push(b"hello").unwrap();
-    u8pool.push(b"world").unwrap();
-
-    let collected: Vec<_> = u8pool.iter().collect();
-    assert_eq!(collected, vec![&b"hello"[..], &b"world"[..]]);
+    assert_eq!(forward, reverse);
 }
