+++ u8pool/tests/stack_tests.rs
@@ -1,6 +1,95 @@
 use u8pool::U8Pool;
 
 #[test]
+fn test_buffer_initialization() {
+    let mut buffer = [0u8; 600];
+    let u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    assert_eq!(u8pool.len(), 0);
+    assert!(u8pool.is_empty());
+}
+
+#[test]
+fn test_bounds_checking_empty_buffer() {
+    let mut buffer = [0u8; 0];
+    assert!(U8Pool::with_default_max_slices(&mut buffer).is_err());
+
+    let mut buffer = [0u8; 600];
+    let u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    assert!(u8pool.get(0).is_none());
+}
+
+#[test]
+fn test_pop_empty_vector() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+    assert_eq!(u8pool.pop(), None); // Should return None
+}
+
+#[test]
+fn test_get_out_of_bounds() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    u8pool.push(b"test").unwrap();
+    assert!(u8pool.get(1).is_none()); // Should return None
+}
+
+#[test]
+fn test_clear_operation() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    u8pool.push(b"hello").unwrap();
+    u8pool.push(b"world").unwrap();
+
+    assert_eq!(u8pool.len(), 2);
+
+    u8pool.clear();
+
+    assert_eq!(u8pool.len(), 0);
+    assert!(u8pool.is_empty());
+}
+
+#[test]
+fn test_pop_operation() {
+    let mut buffer = [0u8; 600];
+    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
+
+    u8pool.push(b"hello").unwrap();
+    u8pool.push(b"world").unwrap();
+
+    let popped = u8pool.pop();
+    assert_eq!(popped, Some(&b"world"[..]));
+    assert_eq!(u8pool.len(), 1);
+
+    let popped = u8pool.pop();
+    assert_eq!(popped, Some(&b"hello"[..]));
+    assert_eq!(u8pool.len(), 0);
+
+    assert_eq!(u8pool.pop(), None);
+}
+
+#[test]
+fn test_custom_max_slices() {
+    let mut buffer = [0u8; 100];
+    let mut u8pool = U8Pool::new(&mut buffer, 3).unwrap();
+
+    u8pool.push(b"test").unwrap();
+    u8pool.push(b"hello").unwrap();
+    u8pool.push(b"world").unwrap();
+
+    // Should fail on 4th slice
+    assert!(u8pool.push(b"fail").is_err());
+
+    assert_eq!(u8pool.get(0).unwrap(), b"test");
+    assert_eq!(u8pool.get(1).unwrap(), b"hello");
+    assert_eq!(u8pool.get(2).unwrap(), b"world");
+    assert_eq!(u8pool.len(), 3);
+}
+
+#[test]
 fn test_stack_push_operations() {
     let mut buffer = [0u8; 600];
     let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
@@ -50,33 +139,6 @@ fn test_stack_push_pop_operations() {
 
     // Test empty pop returns None
     assert_eq!(u8pool.pop(), None);
-}
-
-#[test]
-fn test_stack_interface_doesnt_break_vector_operations() {
-    let mut buffer = [0u8; 600];
-    let mut u8pool = U8Pool::with_default_max_slices(&mut buffer).unwrap();
-
-    // Mix stack and vector operations
-    u8pool.push(b"stack1").unwrap();
-    u8pool.push(b"vector1").unwrap();
-    u8pool.push(b"stack2").unwrap();
-
-    assert_eq!(u8pool.len(), 3);
-    assert_eq!(u8pool.get(0).unwrap(), b"stack1");
-    assert_eq!(u8pool.get(1).unwrap(), b"vector1");
-    assert_eq!(u8pool.get(2).unwrap(), b"stack2");
-
-    // Stack operations still work
-    assert_eq!(u8pool.pop(), Some(&b"stack2"[..]));
-
-    // Vector operations still work
-    assert_eq!(u8pool.get(0).unwrap(), b"stack1");
-    assert_eq!(u8pool.get(1).unwrap(), b"vector1");
-
-    // Iterator still works
-    let collected: Vec<_> = u8pool.iter().collect();
-    assert_eq!(collected, vec![&b"stack1"[..], &b"vector1"[..]]);
 }
 
 #[test]
